add components of Posts and PostEditor #19
